+++ plugin.php
@@ -2,13 +2,13 @@
 $id = "xmum";
 
 // code version; must be changed for all code changes
-$version = "2.0.1";
+$version = "2.2.0";
 
 
 // ilias min and max version; must always reflect the versions that should
 // run with the plugin
 $ilias_min_version = "6.0.0";
-$ilias_max_version = "6.10.999";
+$ilias_max_version = "7.13.999";
 
 // optional, but useful: Add one or more responsible persons and a contact email
 $responsible = "Tobias Goltz";